% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{create_pub_graph}
\alias{create_pub_graph}
\title{Create Publication Graph for Exercise Heart Rate Response}
\usage{
create_pub_graph(
  df = NULL,
  date = NULL,
  lower_bound = NULL,
  upper_bound = NULL
)
}
\arguments{
\item{df}{data fram from \code{\link[=clean_file]{clean_file()}}}

\item{date}{date from \code{\link[=clean_file]{clean_file()}}}

\item{lower_bound}{numeric value from \code{\link[=select_equation]{select_equation()}}}

\item{upper_bound}{numeric value from \code{\link[=select_equation]{select_equation()}}}
}
\value{
Graph with heart rate over time and target intensity zone highlighted
}
\description{
This function will graph the heart rate response over time with target intensity zone highlighted
and will be formatted for publication
}
