% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{enter_age}
\alias{enter_age}
\title{Enter Age}
\usage{
enter_age()
}
\value{
Participant's age
}
\description{
This function allows the user to enter the participant's age. Age is required for calculating maximal heart rate.
}
\details{
Use TryCatch here too. JC
}
\section{Details}{

20240323: Roxygen Skeleton added. BB
}

