% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise_hr_analysis.R
\name{exercise_hr_analysis}
\alias{exercise_hr_analysis}
\title{Exercise HR Analysis Wrapper Function}
\usage{
exercise_hr_analysis(
  output_folder = NULL,
  file = NULL,
  publication_graph = FALSE
)
}
\arguments{
\item{output_folder}{character - A path to where you'd like the image of the graph and the summarized data to be saved}

\item{file}{character - pathway on local computer to unedited Polar HR file to be analyzed}

\item{publication_graph}{Boolean (T/F) - True if you'd like your graph output to be publication quality}
}
\value{
HR graph
}
\description{
This is a master function that uses functions written in support_functions.R to produce
a chart and summarized data
}
