% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{select_bin_size}
\alias{select_bin_size}
\title{Select Bin Size}
\usage{
select_bin_size()
}
\value{
Selected bin size
}
\description{
This function allows users to select 15, 30, 45, or 60 second time bins for dividing the data.
}
\section{Development}{

20240314: Removed df from function parameters
}

