% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{select_equation}
\alias{select_equation}
\title{Select Heart Rate Max Equation}
\usage{
select_equation(lower_bound = NULL, upper_bound = NULL)
}
\arguments{
\item{lower_bound}{numeric value from \code{\link[=select_intensity]{select_intensity()}}}

\item{upper_bound}{numeric value from \code{\link[=select_intensity]{select_intensity()}}}
}
\value{
Maximum heart rate
}
\description{
This function allows the user to choose the appropriate equation for calculating maximum heart rate.
}
\section{Details}{

20240323: Roxygen skeleton added. BB
20240324: Added hr_max variable and changed code to return predicted HRmax. BB
20240326: is the 1st equation right? I see hr_max as 220-age.
}

