% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{select_intensity}
\alias{select_intensity}
\title{Select Intensity}
\usage{
select_intensity()
}
\value{
Selected intensity
}
\description{
This function allows the user to select their target exercise intensity.
}
\section{Details}{

20240319: Return upper bound and lower bound instead of text field. Put text about ACSM selections. JC
20240323: Added in details for light, moderate, vigorous ranges, from ACSM's Guidelines for Exercise Testing and Prescription 11th ed. BB
20240324: Added in if else statements to calculate upper and lower bounds and return these to the user. BB
}

